Updating message.
📂 Archive of 2023 Java Project.
This repository contains files and projects from my CSE studies back in 2022-23. It includes code of a project(tried to create a mobile banking application with 3 other member, i'll try to add them to the collaborator list) and notes that reflect my early journey in programming and computer science. I'm uploading these for reference, documentation, and to track my growth over time.